// apps/resume-app/src/lib/types/PersonalInfo.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ExtraField } from "./ExtraField";

export type PersonalInfo = { fullName: string, title: string | null, email: string | null, phone: string | null, location: string | null, photo: string | null, extras: Array<ExtraField>, };